Create Homebrew Formulae with Descriptions
Until recently, you could call 'brew search' without arguments and brew would return a list of all the formulae on Homebrew. That was useful. I piped that into 'xargs -n 1 brew info' and made this handy reference. It's a little outdated but still useful. If you happen to know of a method for doing the same thing that still works, let me know.